Small Changes and a start with Controllers and Views

+++ RentACar.DataContext/Models/Reservation.cs
@@ -6,7 +6,7 @@ namespace RentACar.DataContext.Models
     public class Reservation
     {
         public int ReservationId { get; set; }
-        public int FinalPrice { get; set; }
+        public float FinalPrice { get; set; }
         public DateTime BeginDate { get; set; }
         public DateTime EndDate { get; set; }
         public int NumberOfDays { get; set; }
