+++ RentACar.ServicesInfrastructure/DTO/ReservationDTO.cs
@@ -6,7 +6,7 @@ namespace RentACar.ServicesInfrastructure.DTO
     public class ReservationDTO
     {
         public int ReservationId { get; set; }
-        public int FinalPrice { get; set; }
+        public float FinalPrice { get; set; }
         public DateTime BeginDate { get; set; }
         public DateTime EndDate { get; set; }
         public int NumberOfDays { get; set; }
